refactoring
extract purchase to package
