+++ backend/receipts/receipt.go
@@ -3,34 +3,28 @@ package receipts
 import (
 	"encoding/json"
 	"go.mongodb.org/mongo-driver/bson/primitive"
+	"receipt_collector/receipts/purchase"
 	"time"
 )
 
-type Purchase struct {
-	Price      int32      `json:"price"`
-	Sum        int32      `json:"sum"`
-	Quantity   float32    `json:"quantity"`
-	Name       string     `json:"name"`
-	Categories []Category `json:"categories"`
-}
-
 type Receipt struct {
-	DateTime             string     `json:"dateTime"`
-	TotalSum             int32      `json:"totalSum"`
-	RetailPlaceAddress   string     `json:"retailPlaceAddress"`
-	UserInn              string     `json:"userInn"`
-	Items                []Purchase `json:"items"`
-	RawData              string     `json:"rawData"`
-	Operator             string     `json:"operator"`
-	Nds18                int32      `json:"nds18"`
-	Nds10                int32      `json:"nds10"`
-	User                 string     `json:"user"`
-	CashTotalSum         int32      `json:"cashTotalSum"`
-	EcashTotalSum        int32      `json:"ecashTotalSum"`
-	FiscalSign           int64      `json:"fiscalSign"`
-	FiscalDocumentNumber int64      `json:"fiscalDocumentNumber"`
+	DateTime             string              `json:"dateTime"`
+	TotalSum             int32               `json:"totalSum"`
+	RetailPlaceAddress   string              `json:"retailPlaceAddress"`
+	UserInn              string              `json:"userInn"`
+	Items                []purchase.Purchase `json:"items"`
+	RawData              string              `json:"rawData"`
+	Operator             string              `json:"operator"`
+	Nds18                int32               `json:"nds18"`
+	Nds10                int32               `json:"nds10"`
+	User                 string              `json:"user"`
+	CashTotalSum         int32               `json:"cashTotalSum"`
+	EcashTotalSum        int32               `json:"ecashTotalSum"`
+	FiscalSign           int64               `json:"fiscalSign"`
+	FiscalDocumentNumber int64               `json:"fiscalDocumentNumber"`
 }
 
+//UsersReceipt is Receipt linked to user.
 type UsersReceipt struct {
 	*Receipt
 	Id                primitive.ObjectID `bson:"_id,omitempty" json:"id"`
@@ -44,6 +38,7 @@ type UsersReceipt struct {
 	Deleted           bool               `json:"deleted" bson:"deleted"`
 }
 
+//RequestStatus - get receipt from nalog.ru API status.
 type RequestStatus string
 
 const (
@@ -53,6 +48,7 @@ const (
 	NotFound  = RequestStatus("not_found")
 )
 
+//ParseReceipt parse API response as Receipt.
 func ParseReceipt(bytes []byte) (Receipt, error) {
 	var receipt map[string]map[string]Receipt
 	err := json.Unmarshal(bytes, &receipt)
@@ -62,15 +58,3 @@ func ParseReceipt(bytes []byte) (Receipt, error) {
 	res := receipt["document"]["receipt"]
 	return res, nil
 }
-
-type Category string
-
-const (
-	Food          Category = "food"
-	Alcohol       Category = "alcohol"
-	Clothes       Category = "clothes"
-	Shoes         Category = "shoes"
-	Medicine      Category = "medicine"
-	HomeAppliance Category = "home_appliance"
-	Entertainment Category = "entertainment"
-)
